fixed so change of vote on same option doesnt send new updates to dat…

+++ backend/src/main/java/com/gruppe2/backend/service/HibernatePollService.java
@@ -225,6 +225,12 @@ public class HibernatePollService implements PollService {
 
             if (poll.getMaxVotesPerUser() == 1) {
                 if (!existingVotes.isEmpty()) {
+                    Vote existingVote = existingVotes.get(0);
+
+                    if (existingVote.getChosenOption().equals(chosenOption)) {
+                        return true;
+                    }
+
                     em.remove(existingVotes.get(0));
                 }
             } else {
